// src/data/profile.d.ts
export declare const profile: {
    name: string;
    englishName: string;
    degree: string;
    title: string;
    address: string;
    about: string[];
    aboutWithLinks: string[];
    social: {
        email: string;
        github: string;
        linkedin: string;
        scholar: string;
    };
    researchInterests: {
        description: string;
        tags: {
            label: string;
            color: string;
        }[];
    };
};
//# sourceMappingURL=profile.d.ts.map